store_monitoring_system/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
kafka-python
redis
opencv-python
torch
ultralytics
sentence-transformers
pydantic
python-dotenv
alembic
pytest